simulation/requirements.txt
kafka-python==2.0.2
redis==4.5.4
